+++ src/main/java/org/truffleruby/language/globals/AliasGlobalVarNode.java
@@ -9,10 +9,8 @@
  */
 package org.truffleruby.language.globals;
 
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import org.truffleruby.language.RubyNode;
-import org.truffleruby.language.control.RaiseException;
 
 public class AliasGlobalVarNode extends RubyNode {
 
@@ -26,17 +24,8 @@ public class AliasGlobalVarNode extends RubyNode {
 
     @Override
     public Object execute(VirtualFrame frame) {
-        checkExisting();
         getContext().getCoreLibrary().getGlobalVariables().alias(oldName, newName);
         return nil();
-    }
-
-    @TruffleBoundary
-    private void checkExisting() {
-        if (getContext().getCoreLibrary().getGlobalVariables().contains(newName)) {
-            // TODO CS 4-Apr-18 This exception is non-standard
-            throw new RaiseException(getContext(), coreExceptions().notImplementedError(String.format("%s is already a global", newName), this), true);
-        }
     }
 
 }