+++ src/main/java/org/truffleruby/language/globals/ReadSimpleGlobalVariableNode.java
@@ -7,27 +7,33 @@ import com.oracle.truffle.api.dsl.Specialization;
 
 public abstract class ReadSimpleGlobalVariableNode extends RubyBaseNode {
 
-    protected final GlobalVariableStorage storage;
+    protected final String name;
 
-    public static ReadSimpleGlobalVariableNode create(GlobalVariableStorage storage) {
-        return ReadSimpleGlobalVariableNodeGen.create(storage);
+    public static ReadSimpleGlobalVariableNode create(String name) {
+        return ReadSimpleGlobalVariableNodeGen.create(name);
     }
 
-    public ReadSimpleGlobalVariableNode(GlobalVariableStorage storage) {
-        this.storage = storage;
+    public ReadSimpleGlobalVariableNode(String name) {
+        this.name = name;
     }
 
     public abstract Object execute();
 
-    @Specialization(assumptions = "storage.getUnchangedAssumption()")
+    @Specialization(assumptions = { "storage.getUnchangedAssumption()", "storage.getValidAssumption()" })
     public Object readConstant(
+            @Cached("getStorage()") GlobalVariableStorage storage,
             @Cached("storage.getValue()") Object value) {
         return value;
     }
 
-    @Specialization
-    public Object read() {
+    @Specialization(assumptions = "storage.getValidAssumption()")
+    public Object read(
+            @Cached("getStorage()") GlobalVariableStorage storage) {
         return storage.getValue();
+    }
+
+    protected GlobalVariableStorage getStorage() {
+        return coreLibrary().getGlobalVariables().getStorage(name);
     }
 
 }