+++ src/main/java/org/truffleruby/language/globals/WriteGlobalVariableNode.java
@@ -22,21 +22,21 @@ import org.truffleruby.language.yield.YieldNode;
 @NodeChild(value = "value")
 public abstract class WriteGlobalVariableNode extends RubyNode {
 
-    private final String name;
+    protected final String name;
 
     public WriteGlobalVariableNode(String name) {
         this.name = name;
     }
 
-    @Specialization(guards = "storage.isSimple()")
+    @Specialization(guards = "storage.isSimple()", assumptions = "storage.getValidAssumption()")
     public Object write(VirtualFrame frame, Object value,
             @Cached("getStorage()") GlobalVariableStorage storage,
-            @Cached("create(storage)") WriteSimpleGlobalVariableNode simpleNode) {
+            @Cached("create(name)") WriteSimpleGlobalVariableNode simpleNode) {
         simpleNode.execute(value);
         return value;
     }
 
-    @Specialization(guards = { "storage.hasHooks()", "arity != 2" })
+    @Specialization(guards = { "storage.hasHooks()", "arity != 2" }, assumptions = "storage.getValidAssumption()")
     public Object writeHooks(VirtualFrame frame, Object value,
                              @Cached("getStorage()") GlobalVariableStorage storage,
                              @Cached("setterArity(storage)") int arity,
@@ -45,7 +45,7 @@ public abstract class WriteGlobalVariableNode extends RubyNode {
         return value;
     }
 
-    @Specialization(guards = { "storage.hasHooks()", "arity == 2" })
+    @Specialization(guards = { "storage.hasHooks()", "arity == 2" }, assumptions = "storage.getValidAssumption()")
     public Object writeHooksWithBinding(VirtualFrame frame, Object value,
             @Cached("getStorage()") GlobalVariableStorage storage,
             @Cached("setterArity(storage)") int arity,