+++ src/main/java/org/truffleruby/language/globals/WriteSimpleGlobalVariableNode.java
@@ -9,22 +9,24 @@ import com.oracle.truffle.api.dsl.Specialization;
 
 public abstract class WriteSimpleGlobalVariableNode extends RubyBaseNode {
 
-    protected final GlobalVariableStorage storage;
+    protected final String name;
     @Child protected ReferenceEqualNode referenceEqualNode = ReferenceEqualNode.create();
     @Child protected WriteBarrierNode writeBarrierNode = WriteBarrierNode.create();
 
-    public static WriteSimpleGlobalVariableNode create(GlobalVariableStorage storage) {
-        return WriteSimpleGlobalVariableNodeGen.create(storage);
+    public static WriteSimpleGlobalVariableNode create(String name) {
+        return WriteSimpleGlobalVariableNodeGen.create(name);
     }
 
-    public WriteSimpleGlobalVariableNode(GlobalVariableStorage storage) {
-        this.storage = storage;
+    public WriteSimpleGlobalVariableNode(String name) {
+        this.name = name;
     }
 
     public abstract Object execute(Object value);
 
-    @Specialization(guards = "referenceEqualNode.executeReferenceEqual(value, previousValue)", assumptions = "storage.getUnchangedAssumption()")
+    @Specialization(guards = "referenceEqualNode.executeReferenceEqual(value, previousValue)",
+                    assumptions = { "storage.getUnchangedAssumption()", "storage.getValidAssumption()" })
     public Object writeTryToKeepConstant(Object value,
+            @Cached("getStorage()") GlobalVariableStorage storage,
             @Cached("storage.getValue()") Object previousValue) {
         // NOTE: we still do the volatile write to get the proper memory barrier,
         // as the global variable could be used as a publication mechanism.
@@ -32,8 +34,10 @@ public abstract class WriteSimpleGlobalVariableNode extends RubyBaseNode {
         return previousValue;
     }
 
-    @Specialization(guards = "storage.isAssumeConstant()", assumptions = "storage.getUnchangedAssumption()")
-    public Object writeAssumeConstant(Object value) {
+    @Specialization(guards = "storage.isAssumeConstant()",
+                    assumptions = { "storage.getUnchangedAssumption()", "storage.getValidAssumption()" })
+    public Object writeAssumeConstant(Object value,
+            @Cached("getStorage()") GlobalVariableStorage storage) {
         if (getContext().getSharedObjects().isSharing()) {
             writeBarrierNode.executeWriteBarrier(value);
         }
@@ -42,13 +46,18 @@ public abstract class WriteSimpleGlobalVariableNode extends RubyBaseNode {
         return value;
     }
 
-    @Specialization(guards = "!storage.isAssumeConstant()", replaces = "writeAssumeConstant")
-    public Object write(Object value) {
+    @Specialization(guards = "!storage.isAssumeConstant()", assumptions = "storage.getValidAssumption()", replaces = "writeAssumeConstant")
+    public Object write(Object value,
+            @Cached("getStorage()") GlobalVariableStorage storage) {
         if (getContext().getSharedObjects().isSharing()) {
             writeBarrierNode.executeWriteBarrier(value);
         }
         storage.setValueInternal(value);
         return value;
+    }
+
+    protected GlobalVariableStorage getStorage() {
+        return coreLibrary().getGlobalVariables().getStorage(name);
     }
 
 }